+++ LCS/Blob/include/Blob/BlobIBuffer.h
@@ -62,7 +62,7 @@ namespace LOFAR {
 	{};
       
       // Get the requested nr of bytes.
-      virtual uint get (void* buffer, uint nbytes) = 0;
+      virtual uint64 get (void* buffer, uint64 nbytes) = 0;
       
       // Get the position in the stream.
       // -1 is returned if the stream is not seekable.
