+++ LCS/Blob/include/Blob/BlobOBufStream.h
@@ -47,7 +47,7 @@ namespace LOFAR {
       virtual ~BlobOBufStream();
       
       // Put the requested nr of bytes.
-      virtual uint put (const void* buffer, uint nbytes);
+      virtual uint64 put (const void* buffer, uint64 nbytes);
       
       // Get the position in the stream.
       // -1 is returned if the stream is not seekable.
